@@ -31,7 +31,7 @@ fun String?.containsCaseInsensitive(other: String?) =
     if (this == null && other == null) {
         true
     } else if (this != null && other != null) {
-        toLowerCase().contains(other.toLowerCase())
+        lowercase().contains(other.lowercase())
     } else {
         false
     }
