@@ -2,8 +2,6 @@ package io.github.gmathi.novellibrary.model.source
 
 import io.github.gmathi.novellibrary.model.other.NovelsPage
 import io.github.gmathi.novellibrary.model.source.filter.FilterList
-import io.github.gmathi.novellibrary.util.lang.awaitSingle
-import rx.Observable
 
 interface CatalogueSource : Source {
 
@@ -18,41 +16,27 @@ interface CatalogueSource : Source {
     val supportsLatest: Boolean
 
     /**
-     * Returns an observable containing a page with a list of novel.
+     * Returns a page with a list of novel.
      *
      * @param page the page number to retrieve.
      */
-    fun fetchPopularNovels(page: Int): Observable<NovelsPage>
+    suspend fun getPopularNovels(page: Int): NovelsPage
 
     /**
-     * [1.x API] Return all the popular novels for this source.
-     */
-    suspend fun getPopularNovels(page: Int): NovelsPage {
-        return fetchPopularNovels(page).awaitSingle()
-    }
-
-    /**
-     * Returns an observable containing a page with a list of novels.
+     * Returns a page with a list of novels.
      *
      * @param page the page number to retrieve.
      * @param query the search query.
      * @param filters the list of filters to apply.
      */
-    fun fetchSearchNovels(page: Int, query: String, filters: FilterList): Observable<NovelsPage>
+    suspend fun getSearchNovels(page: Int, query: String, filters: FilterList = FilterList(emptyList())): NovelsPage
 
     /**
-     * [1.x API] Search and return all the novels that return for that query (search term).
-     */
-    suspend fun getSearchNovels(page: Int, query: String, filters: FilterList = FilterList(emptyList())): NovelsPage {
-        return fetchSearchNovels(page, query, filters).awaitSingle()
-    }
-
-    /**
-     * Returns an observable containing a page with a list of latest novel updates.
+     * Returns a page with a list of latest novel updates.
      *
      * @param page the page number to retrieve.
      */
-    fun fetchLatestUpdates(page: Int): Observable<NovelsPage>
+    suspend fun getLatestUpdates(page: Int): NovelsPage
 
     /**
      * Returns the list of filters for the source.
