@@ -11,6 +11,11 @@ import kotlinx.serialization.json.*
 import uy.kohesive.injekt.injectLazy
 import java.util.Date
 
+/**
+ * Exception thrown when app update operations fail
+ */
+class AppUpdateException(message: String, cause: Throwable? = null) : Exception(message, cause)
+
 internal class AppUpdateGithubApi {
 
     private val networkService: NetworkHelper by injectLazy()
@@ -19,24 +24,32 @@ internal class AppUpdateGithubApi {
 
     private suspend fun getLatestUpdate(): LatestUpdate {
         return withIOContext {
-            networkService.client
-                .newCall(GET("${RELEASES_URL_PREFIX}latest.json"))
-                .await()
-                .parseAs<JsonObject>()
-                .let { parseResponse(it) }
+            try {
+                networkService.client
+                    .newCall(GET("${RELEASES_URL_PREFIX}latest.json"))
+                    .awaitSuccess()
+                    .parseAs<JsonObject>()
+                    .let { parseResponse(it) }
+            } catch (e: Exception) {
+                throw AppUpdateException("Failed to fetch latest update information", e)
+            }
         }
     }
 
     @ExperimentalSerializationApi
     suspend fun checkForUpdates(context: Context): LatestUpdate {
-        val latestUpdate = getLatestUpdate()
-        dataCenter.lastExtCheck = Date().time
+        return try {
+            val latestUpdate = getLatestUpdate()
+            dataCenter.lastExtCheck = Date().time
 
-        if (latestUpdate.versionCode > BuildConfig.VERSION_CODE) {
-            latestUpdate.hasUpdate = true
+            if (latestUpdate.versionCode > BuildConfig.VERSION_CODE) {
+                latestUpdate.hasUpdate = true
+            }
+
+            latestUpdate
+        } catch (e: Exception) {
+            throw AppUpdateException("Failed to check for app updates", e)
         }
-
-        return latestUpdate
     }
 
     private fun parseResponse(json: JsonObject): LatestUpdate {
