@@ -5,6 +5,7 @@ import com.github.salomonbrys.kotson.fromJson
 import com.google.gson.Gson
 import io.github.gmathi.novellibrary.network.NetworkHelper
 import io.github.gmathi.novellibrary.network.WebPageDocumentFetcher
+import kotlinx.coroutines.runBlocking
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.Response
@@ -66,9 +67,22 @@ open class BasePostProxyHelper {
 
     /** Connection used when requesting document. */
     open fun request(url: String): Request = WebPageDocumentFetcher.request(url)
+    
+    // Synchronous methods (original API)
     open fun connect(request: Request): Response = WebPageDocumentFetcher.connect(request)
-    open fun document(response: Response): Document = WebPageDocumentFetcher.document(response)
     open fun string(response: Response): String? = WebPageDocumentFetcher.string(response)
+    open fun document(response: Response): Document = WebPageDocumentFetcher.document(response)
+    
+    // Suspend versions for coroutine support
+    open suspend fun connectAsync(request: Request): Response = WebPageDocumentFetcher.connectAsync(request)
+    open suspend fun stringAsync(response: Response): String? = WebPageDocumentFetcher.stringAsync(response)
+    
+    // Deprecated synchronous versions for backward compatibility
+    @Deprecated("Use connect instead", ReplaceWith("connect(request)"))
+    open fun connectSync(request: Request): Response = WebPageDocumentFetcher.connect(request)
+    
+    @Deprecated("Use string instead", ReplaceWith("string(response)"))
+    open fun stringSync(response: Response): String? = WebPageDocumentFetcher.string(response)
 }
 
 private data class INovelTranslationJson(val chapter: Int, val content: String, val id: Int, val tier: Any?, val tierId: Any?, val title: String?, val volume: Int?, val novel: INovelTranslationJsonNovel)
